# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eBPFsentinel Agent API",
    "description": "REST API for the eBPFsentinel network security agent.\n\nProvides endpoints for managing firewall rules, IPS/IDS policies, rate limiting, L7 filtering, threat intelligence, alerts, and audit logs.\n\nMost endpoints require authentication via JWT Bearer token or API key.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "v0.0.0-dev"
  },
  "paths": {
...
    "/api/v1/conntrack/connections": {
      "get": {
        "tags": [
          "ConnTrack"
        ],
        "summary": "`GET /api/v1/conntrack/connections` - list tracked connections.",
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "Connection list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/conntrack/events": {
      "get": {
        "tags": [
          "ConnTrack"
        ],
        "summary": "`GET /api/v1/conntrack/events` - Server-Sent Events stream of\nconntrack lifecycle events (new / update / destroy).",
        "description": "The poller diffs conntrack snapshots every 2 s and\npushes changes into a broadcast channel. Each SSE client receives\na copy. Lagged clients silently skip missed events.",
        "operationId": "conntrack_events",
        "responses": {
          "200": {
            "description": "SSE stream of `event: new`, `event: update` and `event: destroy` frames. Each frame's `data` is one JSON object carrying `event_type` and the `connection` it happened to; `:keepalive` is sent every 15 s.",
            "content": {
              "text/event-stream": {}
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Conntrack event stream not enabled, which is also what a kernel built without `CONFIG_NF_CONNTRACK_PROCFS` answers, since the poller has nothing to read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/conntrack/flush": {
      "post": {
        "tags": [
          "ConnTrack"
        ],
        "summary": "`POST /api/v1/conntrack/flush` - flush all tracked connections.",
        "operationId": "flush_connections",
        "responses": {
          "200": {
            "description": "Connections flushed"
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
    "/api/v1/conntrack/status": {
      "get": {
        "tags": [
          "ConnTrack"
        ],
        "summary": "`GET /api/v1/conntrack/status` - connection tracking status.",
        "operationId": "conntrack_status",
        "responses": {
          "200": {
            "description": "Conntrack status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnTrackStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          },
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnTrackStatusResponse": {
        "type": "object",
        "required": [
          "enabled",
          "connection_count",
          "max_connections"
        ],
        "properties": {
          "connection_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "enabled": {
            "type": "boolean"
          },
          "max_connections": {
            "type": "integer",
            "format": "int64",
            "description": "Conntrack table capacity - the maximum number of flows the\nagent tracks before eviction (the BPF connection-table size).",
            "minimum": 0
          }
        }
      },
      "ConnectionResponse": {
        "type": "object",
        "required": [
          "src_ip",
          "dst_ip",
          "src_port",
          "dst_port",
          "protocol",
          "state",
          "packets_fwd",
          "packets_rev",
          "bytes_fwd",
          "bytes_rev"
        ],
        "properties": {
          "bytes_fwd": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "bytes_rev": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "dst_ip": {
            "type": "string"
          },
          "dst_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "packets_fwd": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "packets_rev": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "protocol": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "src_ip": {
            "type": "string"
          },
          "src_port": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "state": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `RULE_NOT_FOUND`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key configured in agent"
      },
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained via OIDC or static configuration"
      }
    }
  },
...
}